ISSUE-7: less responsibility for API and survey provider protocol
