+++ src/survey-list-app/Model/API.swift
@@ -9,63 +9,18 @@
 import Foundation
 import Alamofire
 
-// API communicates with the survey API, does autorization and retrieves the list of surveys
+// API provides access to the API and does auth
 class API {
+    // afSessionManager for auth providers
+    let afSessionManager = Alamofire.SessionManager(configuration: URLSessionConfiguration.default)
+    
     // see survey-list.xcconfig
-    private let baseUrl = "https://\(Bundle.main.object(forInfoDictionaryKey:"API_BASE_URL") as! String)/"
+    let baseUrl = "https://\(Bundle.main.object(forInfoDictionaryKey:"API_BASE_URL") as! String)/"
     private let authUser = Bundle.main.object(forInfoDictionaryKey:"API_AUTH_USER") as! String
     private let authPass = Bundle.main.object(forInfoDictionaryKey:"API_AUTH_PASS") as! String
     
     // bearer token for auth
     private var bearer: Bearer?
-    private let afSessionManager = Alamofire.SessionManager(configuration: URLSessionConfiguration.default)
-    
-    // getSurveys retrieves surveys from the API and handles authentication if needed; results can be paged optionally
-    func getSurveys(page: Int? = nil, perPage: Int? = nil, completion: ((Error?, [Survey]?) -> ())? = nil) {
-        authIfNeeded() { error, bearer in
-            guard let bearer = bearer, error == nil else {
-                completion?(error, nil)
-                return
-            }
-            
-            var params = [ "access_token": bearer.accessToken ]
-            if let page = page {
-                params["page"] = String(page)
-            }
-            if let perPage = page {
-                params["per_page"] = String(perPage)
-            }
-            
-            self.afSessionManager.request("\(self.baseUrl)surveys.json", method: .get, parameters: params)
-            .responseData { response in
-                if let error = response.result.error {
-                    // treat alamofire error as network error
-                    completion?(APIError.network(error), nil)
-                    return
-                }
-                guard let status = response.response?.statusCode, status == 200 else {
-                    // treat response format problems as backend error
-                    completion?(APIError.backend, nil)
-                    return
-                }
-                
-                do {
-                    guard let result = response.result.value else {
-                        throw APIError.backend
-                    }
-                    // api responds sometimes with null and sometimes with [] when page is empty.
-                    // so we check for null here. optional JSONDecoder [Survey]?.self doesn't suffice
-                    if "null" == String(data: result, encoding: .utf8) {
-                        completion?(nil, [])
-                    } else {
-                        completion?(nil, try JSONDecoder().decode([Survey].self, from: result))
-                    }
-                } catch {
-                    completion?(error, nil)
-                }
-            }
-        }
-    }
     
     // authIfNeeded authorizes API if no token is available or if it expires soon.
     func authIfNeeded(completion: ((Error?, Bearer?) -> ())? = nil) {
